Authentication (#1)
* Added signup endpoint

* Fixed bugs

* Added login feature

* Added basic user profile

* Added validate token middleware

* Added authorisation for post upload

* Completed authorisation for upload service

* Added refresh token endpoint

* Added authorisation for getting podcast

* Added readme

* Updated image link

* Updated image link

* Refinied readme

* Completed authorisation for podcast CRUD

* Added authorisation for all podcast endpoints

+++ functions/src/services/upload.ts
@@ -8,18 +8,19 @@ import {
   ImagesStorage,
 } from "../database/db";
 
-const postOneUpload = async (filetype: FileType, filepath: string) => {
+const postOneUpload = async (upload: Upload) => {
   try {
     const data: Upload = {
-      filetype: filetype,
-      filepath: filepath.split("_").slice(-1)[0],
+      userId: upload.userId,
+      filetype: upload.filetype,
+      filepath: upload.filepath.split("_").slice(-1)[0],
     };
 
-    if (filetype === FileType.PODCAST) {
-      const duration = getAudioDurationInSeconds(filepath);
+    if (data.filetype === FileType.PODCAST) {
+      const duration = getAudioDurationInSeconds(upload.filepath);
 
       // Upload to cloud storage
-      await PodcastsStorage.upload(filepath, {
+      await PodcastsStorage.upload(upload.filepath, {
         destination: data.filepath,
       });
 
@@ -31,9 +32,9 @@ const postOneUpload = async (filetype: FileType, filepath: string) => {
       });
 
       return { podcastUploadId: res.id };
-    } else if (filetype === FileType.IMAGE) {
+    } else if (data.filetype === FileType.IMAGE) {
       // Upload to cloud storage
-      await ImagesStorage.upload(filepath, {
+      await ImagesStorage.upload(upload.filepath, {
         destination: data.filepath,
       });
 
@@ -50,43 +51,47 @@ const postOneUpload = async (filetype: FileType, filepath: string) => {
   } catch (err: any) {
     throw { status: err?.status || 500, message: err?.message || err };
   } finally {
-    fs.unlinkSync(filepath);
+    fs.unlinkSync(upload.filepath);
   }
 };
 
-const updateOneUpload = async (
-  filetype: FileType,
-  uploadId: string,
-  updatedFilepath: string
-) => {
+const updateOneUpload = async (uploadId: string, updatedUpload: Upload) => {
   try {
     const updatedData: Upload = {
-      filetype: filetype,
-      filepath: updatedFilepath.split("_").slice(-1)[0],
+      userId: updatedUpload.userId,
+      filetype: updatedUpload.filetype,
+      filepath: updatedUpload.filepath.split("_").slice(-1)[0],
     };
 
-    if (filetype === FileType.PODCAST) {
-      const duration = getAudioDurationInSeconds(updatedFilepath);
+    const res = await UploadsCollection.doc(uploadId).get();
+    if (!res.exists) {
+      throw { status: 404, message: `Upload id ${uploadId} not found.` };
+    }
 
-      const res = await UploadsCollection.doc(uploadId).get();
-      if (!res.exists) {
-        throw { status: 404, message: `Upload id ${uploadId} not found.` };
-      }
+    const oldFilepath = res.data()?.filepath;
+    const oldFiletype = res.data()?.filetype;
+    if (oldFilepath == undefined || oldFiletype != updatedData.filetype) {
+      throw { status: 404, message: `Upload id ${uploadId} not found.` };
+    }
 
-      const oldFilepath = res.data()?.filepath;
-      const oldFiletype = res.data()?.filetype;
-      if (oldFilepath == undefined || oldFiletype != filetype) {
-        throw { status: 404, message: `Upload id ${uploadId} not found.` };
-      }
+    if (res.data()?.userId != updatedData.userId) {
+      throw {
+        status: 403,
+        message: `You do not have access to upload ${uploadId} not found.`,
+      };
+    }
+
+    if (updatedData.filetype === FileType.PODCAST) {
+      const duration = getAudioDurationInSeconds(updatedUpload.filepath);
 
       // Upload new file to cloud storage
-      await PodcastsStorage.upload(updatedFilepath, {
+      await PodcastsStorage.upload(updatedUpload.filepath, {
         destination: updatedData.filepath,
       });
 
       // Update in firestore
       updatedData.durationInMinutes = Math.round(await duration);
-      const data = await UploadsCollection.doc(uploadId)
+      await UploadsCollection.doc(uploadId)
         .update(updatedData)
         .catch(async (err: any) => {
           await PodcastsStorage.file(updatedData.filepath).delete();
@@ -96,26 +101,15 @@ const updateOneUpload = async (
       // Delete old file from cloud storage
       await PodcastsStorage.file(oldFilepath).delete();
 
-      return data;
-    } else if (filetype === FileType.IMAGE) {
-      const res = await UploadsCollection.doc(uploadId).get();
-      if (!res.exists) {
-        throw { status: 404, message: `Upload id ${uploadId} not found.` };
-      }
-
-      const oldFilepath = res.data()?.filepath;
-      const oldFiletype = res.data()?.filetype;
-      if (oldFilepath == undefined || oldFiletype != filetype) {
-        throw { status: 404, message: `Upload id ${uploadId} not found.` };
-      }
-
+      return { status: "OK", message: "Your upload has been updated." };
+    } else if (updatedData.filetype === FileType.IMAGE) {
       // Upload new file to cloud storage
-      await ImagesStorage.upload(updatedFilepath, {
+      await ImagesStorage.upload(updatedUpload.filepath, {
         destination: updatedData.filepath,
       });
 
       // Update in firestore
-      const data = await UploadsCollection.doc(uploadId)
+      await UploadsCollection.doc(uploadId)
         .update(updatedData)
         .catch(async (err: any) => {
           await ImagesStorage.file(updatedData.filepath).delete();
@@ -125,18 +119,18 @@ const updateOneUpload = async (
       // Delete old file from cloud storage
       await ImagesStorage.file(oldFilepath).delete();
 
-      return data;
+      return { status: "OK", message: "Your upload has been updated." };
     } else {
       throw { status: 500, message: "Filetype not compatible." };
     }
   } catch (err: any) {
     throw { status: err?.status || 500, message: err?.message || err };
   } finally {
-    fs.unlinkSync(updatedFilepath);
+    fs.unlinkSync(updatedUpload.filepath);
   }
 };
 
-const deleteOneUpload = async (uploadId: string) => {
+const deleteOneUpload = async (uploadId: string, userId: string) => {
   try {
     const res = await UploadsCollection.doc(uploadId).get();
     if (!res.exists) {
@@ -148,12 +142,20 @@ const deleteOneUpload = async (uploadId: string) => {
       throw { status: 404, message: `Upload id ${uploadId} not found.` };
     }
 
+    if (res.data()?.userId != userId) {
+      throw {
+        status: 403,
+        message: `You do not have access to upload ${uploadId} not found.`,
+      };
+    }
+
     // Delete from cloud storage
     await PodcastsStorage.file(filepath).delete();
 
     // Delete from firestore
-    const data = await UploadsCollection.doc(uploadId).delete();
-    return data;
+    await UploadsCollection.doc(uploadId).delete();
+
+    return { status: "OK", message: "Your upload has been deleted." };
   } catch (err: any) {
     throw { status: err?.status || 500, message: err?.message || err };
   }
